Disable youtube recommendations at end of video
Normally, the embedded youtube player will show recommended videos based
on the one played.
This isn't however desired if we want the reader to submerge into a
long, distraction-free reading/viewing experience, so we've disabled the
video recommendations for our projects. I even think this would make a
good default for AESOP.