Fade in quote element at 90% of page height (instead of when element's bottom border is in viewport)
When not using parallax, we discovered that longer quotes would appear
too late. Users would scroll, see a whitespace gap at the bottom of the
page, scroll further, and suddenly the quote would fade in.
For longer quotes, we found that replacing the fade-in policy
"bottom-in-view" with "90%" works reasonably well for both, one-line
quotes and longer quotes (7-8 lines)