Multiple Tile Vendors
Add tiles from Stamen, OSM and more sources. Creates new cmb in
edit.php screen with select that connects to conditional statements in
both component-map.php files (tiles show in backend editor and
front-end display!!) with attribution to sources.

- Maintains backwards compatibility by including Mapbox as a provider
and defaulting to it, as well as including an else statement to always
fallback to the original statement even if custom meta is undefined.
Sloppy, not too much bloat and effective.
- I know it’s less than ideal to add another custom meta box, I was
originally planning to add this as another shortcode element in the
composer, but due to the way Aesop loads the tiles saving as custom
meta was important.