Fix: Youtube embed in IE ignores z-index values
In Internet Explorer 11 (most probably 10 and below, too), when
embedding a Youtube video which relies on the flash player to be played,
we noticed that the player object stays on top of all other objects:
header bar and even the IE11 scroll bar when the width is set to 100%.

This seems to be a well-known problem:
http://stackoverflow.com/questions/9074365/youtube-video-embedded-via-iframe-ignoring-z-index

Apparently, setting wmode="transparent" on the iframe (as it already is)
isn't enough.
By providing the URL parameter "?wmode=transparent", the flash object
(<embed> element) in the iframe will get the parameter
wmode="transparent", which solves the issue.