fix type of video block attributes
the shortcode expects these attributes to be strings, not
booleans. Without the fix, the values of attributes which were
default 'on' in the shortcode (e.g. autoplay) could not be switched
of after they were once changed because the default in the shortcode
overrides the value from the block which was not set.
Using 'string' as the type, if switched off, the option in the shortcode
is set und thus the default is not applied.